+++ dotnet/Base/Database/Configuration.Tests/TestDatabaseServices.cs
@@ -6,14 +6,14 @@
 
 namespace Allors.Database.Configuration
 {
-    using System.IO;
     using Database.Derivations;
     using Derivations.Default;
     using Domain;
+    using Microsoft.Extensions.Configuration;
 
     public class TestDatabaseServices : DatabaseServices
     {
-        public TestDatabaseServices(Engine engine, DirectoryInfo dataPath = null) : base(engine, dataPath) { }
+        public TestDatabaseServices(Engine engine, IConfiguration configuration = null) : base(engine, configuration) { }
 
         protected override IPasswordHasher CreatePasswordHasher() => new TestPasswordHasher();