+++ dotnet/Base/Database/Configuration/Custom/Database/DatabaseServices.cs
@@ -7,12 +7,12 @@
 namespace Allors.Database.Configuration
 {
     using System;
-    using System.IO;
     using Data;
     using Database.Derivations;
     using Derivations.Default;
     using Domain;
     using Meta;
+    using Microsoft.Extensions.Configuration;
     using Ranges;
     using Services;
 
@@ -60,13 +60,13 @@ namespace Allors.Database.Configuration
 
         private IWorkspaceMask workspaceMask;
 
-        protected DatabaseServices(Engine engine, DirectoryInfo dataPath = null)
+        private readonly IConfiguration configuration;
+
+        protected DatabaseServices(Engine engine, IConfiguration configuration = null)
         {
             this.Engine = engine;
-            this.DataPath = dataPath;
+            this.configuration = configuration;
         }
-
-        protected DirectoryInfo DataPath { get; }
 
         internal IDatabase Database { get; private set; }
 
@@ -106,7 +106,7 @@ namespace Allors.Database.Configuration
                 // Base
                 { } type when type == typeof(ISingletonId) => (T)(this.singletonId ??= new SingletonId()),
                 { } type when type == typeof(IMailer) => (T)(this.mailer ??= new MailKitMailer()),
-                { } type when type == typeof(IMediaContentStorage) => (T)(this.mediaContentStorage ??= new FileMediaContentStorage(this.DataPath ?? new DirectoryInfo("media"))),
+                { } type when type == typeof(IMediaContentStorage) => (T)(this.mediaContentStorage ??= new FileMediaContentStorage(FileMediaContentStorage.ResolveDirectory(this.configuration))),
                 { } type when type == typeof(IBarcodeGenerator) => (T)(this.barcodeGenerator ??= new ZXingBarcodeGenerator()),
                 { } type when type == typeof(ITemplateObjectCache) => (T)(this.templateObjectCache ??= new TemplateObjectCache()),
                 _ => throw new NotSupportedException($"Service {typeof(T)} not supported")